customized alerts for stations with slashes
For commuter rail alerts, specific stations' names are interpolated into
the description. For example, if the description is:

> Affected trips: Newburyport Train 180 (9:25 PM from Newburyport)

But the station someone is interested in is Chelsea, then the
description in the alert will be updated to look like this:

> Affected trips: Newburyport Train 180 (22:17 pm from Chelsea)

This functionality was not working with stations with slashes in them
(or more generally, any station where the station `name` and station
`id` are the same). The fix is to match against the station `id` instead
of the `name` and only use the `name` in the description text.