do not show same alert in digest multiple times
If an alert spans multiple date groups (upcoming_weekend,
upcoming_week, next_weekend, future), then only show it once in the
digest email instead of showing it multiple times.

However, if the alert is recurring (ie `alert.active_period` is a list
with multiple members) then show it in each section that it occurs in.