show the alert id for notifications for admin
When an admin looks at "Received Notifications" for a user, show the
alert id in addition to the phone number, email, and alert text.

The alert id is needed in order to use the "Diagnostics" feature, so
including it in the text visible to admins will make it easier for them
to figure out why a certain alert was sent to a user.