do not send estimated alerts immediately (#468)
When an alert's duration is `estimated`, then do not send the first
notification immediately. Instead, send it `A` minutes before the start
of the subscription, where `A` is defined using the formula below:

```
S = advanced notice requirement by severity, in minutes
60 for minor, 120 for moderate, 240 for severe (calculated by Concierge)
D = the projected duration in minutes (as defined in the feed)
A = the minimum of the alert duration D OR severity based advance notice S
```

The code previously used this formula for sending followup notifications.
The only change made in this commit it that it applies the above logic
to *all* notifications, not only followup notifications.

There are other issues with the `estimated` alert sending logic which are
not addressed in this commit. For a summary, see
[here](https://app.asana.com/0/415342363785198/495798504989716/f).