refactor build_subscription_update_transaction (#478)
Remove the function `build_subscription_update_transaction/3` from a
bunch of different `*Mapper` modules and put it in the `Mapper` module
and also add a test. The DRYs up the code and should fix an issue where
CodeCov wasn't recognizing test coverage because the controller tests
that test this functionality are in a different application.

While doing this, I realized there are no other tests for `Mapper`, so I
created a tech debt [ticket](https://app.asana.com/0/415342363785198/502793352397248/f)
to fix that.