password format error message (#495)
When a user types a password with the incorrect format on a modern
browser, the browser will reject it without submitting it to the server
due to the `pattern` attribute.

The browser will then display an error message that includes the `title`
attribute. When the `title` attribute it missing, the error message
may look weird.

This commit adds that attribute, giving the user a more useful error
message.