support bus stops with parent stations (#496)
Sometimes there are elevator delays at bus stops which are part of
subway stations. In the API this is represented by a parent
relationship. However, the code that looks for parents wasn't looking at
busses.

This commit changes the API to look for parent stations for all modes of
transport, updates the tests, and adds some constants to a list of
constants that were causing `:erlang.binary_to_existing_atom` to fail.