dialyzer: use map for Guardian.Plug.sign_in/4
In versions of Guardian less than 1.0, the `sign_in/4` function accepts
either a map or a keyword list as the last argument, but the typespec
says it only supports a map.

See https://github.com/ueberauth/guardian/blob/v0.14.5/lib/guardian/plug.ex#L94

To stop Dialyzer from complaining, in all instances where `sign_in/4` is
called, use a map instead of a keyword list for the last argument.

(This issue was fixed in version 1.0, but 1.0 is not backwards compatible
with 0.14, so upgrading isn't an easy way to fix the problem.)