account for time zone in active period
When converting a Unix timestamp into a `DateTime` for use in
`AlertProcessor.ActivePeriodFilter`, convert it to time zone
`America/New_York`. Otherwise, the comparison will be done on the time
in UTC, which means that alerts won't always match subscriptions.

As part of this PR, refactor the `DateTimeHelper` module to use a time
zone constant instead of repeating the string `"America/New_York"`.