do not ignore priv (#535)
Do not ignore the `priv/` directory in `.dockerignore`. Ignoring this
directory was causing the migrations to not be included in the build.

Also remove the leading `/` from each line. The documentation says that
it should work the same with or without the slash, but in some cases it
doesn't work with the slash (https://stackoverflow.com/questions/36295418/dockerignore-mentioned-files-are-not-ignored)