remove amber alerts code
As far as I can tell, the `amber_alert_opt_in` field is only used on the
user account page and not in any alerts matching logic.

This commit removes that column and any logic around it. (Users could
still receive Amber alerts if they choose to sign up for moderate severity
alerts, but that logic is unrelated.)