filter estimated notifications time zones (#547)
The code that filters whether a notification with estimated duration
certainty is sent today or tomorrow compares the end time of the
subscription to the current time. However it was doing these comparisons
in different time zones. This resulted in estimated alerts not being
sent when they should have been.