add user_id to notification struct (#564)
In `AlertProcessor.NotificationBuilder.build_notifications/1`, include
the `user_id`. The `Notification` struct is expected to have a `user_id`
so this makes it more predictable when used later in the code.

This also fixes a bug in `AlertProcessor.HoldingQueue.enqueue/2`.
It removes duplicates with the following code: `Enum.uniq_by(newstate, &
{&1.user_id, &1.alert_id, &1.send_after})`. Since `user_id` wasn't being
set, it was `nil`, and so it would always match. Now that it's set, the
uniqueness code will work correctly.