informed entities filter refactor (#574)
Add some short circuiting logic to the informed entities filter. I
didn't see that much of a performance change with this code, but that's
expected since we don't have many informed entities in the example.
In the case where there are a lot of them, or the subscription has a lot
of them, this will have a bigger effect.