remove seed file (#577)
One reason that tests fail locally but work in CI is that running `mix
ecto.reset` runs `seeds.exs` in development and test environment. This
contaminates the test DB.

Since we're not using these users for anything, I got rid of the seeds
entirely. If we ever add them back, we should make sure to do it
conditionally on `Mix.env == :dev`