V2 Edit subscription frontend
Why:

* V2 needs a page where users can edit their subscription details.
* Asana link: https://app.asana.com/0/529741067494252/570673981108064

This change addresses the need by:

* Editing `DaySelectHelper` to also accept days as atoms since that's
how they come out of the DB.
* Adding `Trip.find_by_id/1` to load a trip and it's subscriptions with
a given trip id. Depending on the authorization strategy we've decided
to use on this project, we might want to edit this function to also
accept a `user_id` and only return the trip with a valid `user_id`.
* Editing V2.TripController edit action to pull a trip using
`Trip.find_by_id/1`.
* Adding basic HTML to v2 trip edit page:
`templates/v2/trip/edit.html.eex`.
* Adding fallback controller to handle 404s.