support all days of week in notification builder
The function `DateTimeHelper.determine_relevant_day_of_week` is used in
building notifications. Previously it only worked with the values
`[:weekday, :saturday, :sunday]`. This commit allows it to work with any
day of the week (`:monday`, `:tuesday`, etc.).

This is necessary because in v2, users will be able to select a subset
of weekdays.