support days for matching alerts and subscriptions
The function `AlertProcessor.Model.Subscription.timeframe_map/1` is used
when matching alerts and subscriptions. It takes a subscription and
returns a data structure of days and times where the subscription should
match an alert.

This commit allows a subscription to contain individual days of the week
(eg `:monday`, `:tuesday`). Previously it only worked with `[:weekday,
:saturday, :sunday]`.

This is necessary because in v2, users will be able to select a subset of
weekdays.