Add basic trip update functionality
Why:

* We want users to be able to update their trip preferences. We want to
allow them to edit the alert time, days, and trip times. This commit
includes changes to allow users to edit all of these attributes except
for "alert time". Also note that for the "trip update" functionality to
be considered as fully functional we need to make sure a trip's
subscriptions are updated whenever a trip is updated. Support for this
will be added in an upcoming commit.
* Asana link: https://app.asana.com/0/529741067494252/589640936873692

This change addresses the need by:

* Adding `Trip.update/2` to allow us to easily trigger trip updates from
the Trip controller update action. This function needs to be updated in
an upcoming commit for it to trigger subscription updates whenever a
trip is updated.
* Editing the trip controller update action to use the new function
mentioned above.
* Editing the trip edit template for compatibility with the changes made
to the trip controller.
* Editing the DaySelectHelper module for compatibility with target DB
column.
* Editing the trip factory to be able to create trips by calling
`insert(:trip)` within tests. This allows us to not have to create a
user and pass it to the insert function.
* Edit tests that were using `insert(:trip, %{user_id: user_id}` to be
compatible with the changes mentioned above this one.