Add alert_time_difference_in_minutes column
Why:

* Currently we don't have a column in the trips table to store the
desired notification time for a return trip. Instead of creating a new
return_notification_time column for round-trips, we want to remove the
existing notification_time column and replace it with an
alert_time_difference_in_minutes column that would work for both one-way
trips and round-trips. Note that the design for the edit
trips/subscription page allows users to edit this value.
* Asana link: https://app.asana.com/0/529741067494252/599283856113910

This change addresses the need by:

* Adding migration to add `alert_time_difference_in_minutes` column and
remove `notification_time` from trips table. Note that this migrations
also removes a function and trigger that was there to keep trips and
subscriptions in-sync. In an upcoming commit we'll introduce changes to
handle this on the Elixir side.
* Editing `Trip` model and `TripTest` module to use
`alert_time_difference_in_minutes` instead of `notification_time`.
* Editing `Factory` module's `trip_factory/0` by removing
`notification_time` and adding `alert_time_difference_in_minutes`
* Editing various tests that previously relied on the
`notification_time` column in the trips table.