AlertParser supports valid TranslatedString type
Why:

* The feed we're currently working with doesn't correctly format
`TranslatedString` types. In the near future they will. So we want to be
ready for this when the feed starts formatting this data per the
GTFS-realtime specification. This commit extends
`AlertParser.parse_translation/1` for it to be able to parse valid
`TranslatedString` types as well as the incorrect format the feed has
currently generates. For examples of the correct and incorrect format
take a look at the comments in the Asana ticket below.
* Asana link: https://app.asana.com/0/529741067494252/605854572861571