Edit InformedEntityFilter per v2 requirements
Why:

* In what we're calling v2, the `InformedEntityFilter` module should not
rely on a subscription's informed_entities, because in v2 subscriptions
do not have informed_entities. This commit updates the
`InformedEntityFilter` module accordingly.
* Asana link: https://app.asana.com/0/529741067494252/598093528009268

This change addresses the need by:

* Editing `InformedEntityFilter` as described above.
* Fixing all the tests that broke after the changes mentioned above.
This is needed because all these tests were previously relying on
a subscription's informed_entities to work, but the subscription's in the
test were not set up as needed per the updates to the
`InformedEntityFilter` module.