Fix bus trip create bug
Why:

* Currently we can't create bus trips because of the validation that we
recently added to prevent users from creating trips with the same
origins and destinations. More specifically this is because the origin
and destination for a bus trip comes in as two empty strings.
* Asana link: https://app.asana.com/0/529741067494252/640514815408664

This change addresses the need by:

* Adding `:scrub_params` plug to remove empty strings on
V2.TripController's leg action `"trip"` params.
* Editing V2.TripController's leg action to only error and notify user
that origins can't equal destinations if a trip's origin is not nil.