Fix bug: repeated notifications
Why:

* Notifications are being sent twice. This is unwanted behavior. The
cause of this is the `NotificationBuilder` which is doubling up
notifications for some reason. This seems to be related to a module
called `HoldingQueue`, which doesn't exist anymore. The
`NotificationBuilder` is currently building notifications for each of
the given alert's active periods for today, tomorrow, and the day after
tomorrow. When I reproduced the bug locally I noticed that the same
notification is coming back for "today's notifications" and "tomorrow's
notifications". Because the `HoldingQueue` doesn't exist anymore, two
notifications are being sent out immediately. This module also seems to
do some filtering but not sure why. To be honest, I'm not sure what was
the motivation for scheduling notifications three day's in advance but
moving forward it seems more reasonable to have a module called
`NotificationBuilder` to be in charge of building notifications for a
given user-alert match and for filtering of subscriptions to be done
elsewhere. This module is currently also setting the `send_after` value
on notifications, which was only relevant when the `HoldingQueue` was
still in use. Also note that the filtering done in `NotificationBuilder`
seems to be handled by the `NotificationWindowFilter` and the
`ActivePeriodFilter` which the subscriptions go through before passing
through `NotificationBuilder`.
* Asana link: https://app.asana.com/0/529741067494252/640392812479504

This change addresses the need by:

* Editing `NotificationBuilder` to build a single notification for each
alert that is passed to it as an argument. Before, it was building
notifications that should be sent today, tomorrow, and the day after
tomorrow for each of the alert's active periods, which was generating
more notifications than desired.
* Editing `NotificationBuilder` to not do any filtering.
* Editing `NotificationBuilder` to not set the `send_after` field as
it's not needed now that the `HoldingQueue` doesn't exist anymore.