Fix bug: station feature alerts
Why:

* Station feature alerts (accessibility alerts) weren't triggering
notifications. This is unwanted behavior. The culprit is the
`NotificationWindowFilter`. More specifically, the culprit is the
`time_within_notification_window?/2` function in this module.
* Asana link: https://app.asana.com/0/529741067494252/640514815408666

This change addresses the need by:

* Editing `NotificationWindowFilter.time_within_notification_window?/2`
to always return true for accessibility-type subscriptions. This is
because users that sign up for accessibility subscriptions want to be
notified of matching alerts regardless of the current time. In other
words, these type of subscriptions don't have a notification window.