Fix intermittent test failure
Why:

* There is a test in `NotificationWorkerTest` that fails intermittently
on CI.
* Asana link: https://app.asana.com/0/529741067494252/569662205566961

This change addresses the need by:

* Starting `NotificationWoerker` with different names in
`NotificationWorkerTest`. Not 100% this is the only reason why this
fails intermittently. I've seen it fail because it can't start the
NotificationWorker with the same name as it gets started with on other
tests within that file because "it was already started". Some times it
seems to also fail because of an assert_received timeout.