Fix intermittent NotificationWorkerTest failure
Why:

* There is a test in `NotificationWorkerTest` that fails intermittently
and is causing us to manually rebuild on CI. The root of the problem is
a race condition between the two `NotificationWorker` processes started
in `AlertProcessor.Supervisor`'s `init/1` and the `NotificationWorker`
started in the test. These three processes are constantly attempting to
pop off of the `SendingQueue`. In the test we attempt to assert that a
specific `NotificationWorker` process, out of the three that are running
at that point, gets the notification enqueued in the test and processes
it. When this doesn't happen, the test fails.
* Asana link: https://app.asana.com/0/529741067494252/569662205566961

This change addresses the need by:

* Editing the apps/alert_processor/config/test.exs to start 0
NotificationWorker processes in test env. These are not needed when in
test env. This gets rid of the race condition because we no longer have
these two processes trying to process `SendingQueue` elements.
* Editing `NotificationWorkerTest` tests to not start the `SendingQueue`
because it's already started by `AlertProcessor.Supervisor`.
* Editing `NotificatioWorkerTest` to use `start_supervised/1` to start
the `NotificationWorker` process required for each of the tests in that
file. `start_supervised/1` starts a child process under the test
supervisor. This child process, `NotificationWorker` in this scenario,
is guaranteed to exit before the next test starts.
* Editing `NotificationWorkerTest` by removing the `:timer.sleep/1`
calls that are not needed.