Remove excessive API calls
Why:

* Recently, Paul Swartz reported that the alerts concierge API key was
using the MBTA API almost as much as the website. After looking into it
a bit an API usage report for the alerts concierge key shows that 99.3%
(724,193) of the API calls in the last 24 hours were made to the
`/schedules` endpoint. This commit attempts to reduce the amount of API
calls to the `/schedules` endpoint.
* Asana link: https://app.asana.com/0/529741067494252/637596559275142

This change addresses the need by:

* Editing the `AlertParser` to not make calls to the `/schedules`
endpoint for every single alert informed_entity of route_type 2. The
`AlertParser` was hitting the `/schedules` endpoint to add the trip's
"schedule" to commuter rail alerts. This is not necessary because these
schedule values are not being used in any way. Maybe they were used in
the past and once the application stopped relying on those values the
`AlertParser` was not updated?