Fix ServiceInfoCache HTTPoison error
Why:

* On alerts-dev (staging environment) we noticed that the
`ServiceInfoCache` process was terminating with the following error:

```
2018-05-03T14:22:52.875 [error] node=alerts_concierge@127.0.0.1 Task
** (MatchError) no match of right hand side value: {:error,
%HTTPoison.Error{id: nil, reason: :connect_timeout}}
    (alert_processor) lib/service_info/service_info_cache.ex:485:
AlertProcessor.ServiceInfoCache.do_fetch_service_info/1
    (elixir) lib/task/supervised.ex:88: Task.Supervised.do_apply/2
    (elixir) lib/task/supervised.ex:38: Task.Supervised.reply/5
    (stdlib) proc_lib.erl:247: :proc_lib.init_p_do_apply/3
Function: #Function<44.118878969/0 in
AlertProcessor.ServiceInfoCache.fetch_parallel/1>
```

The error seems to be related to a bug in Hackney that was fixed in
Hackney 1.10.1 (https://github.com/edgurgel/httpoison/issues/291). There
was another bug that was not related but that was fixed in Hackney
1.12.0 (https://github.com/edgurgel/httpoison/issues/310). Only
mentioning the latter if you're wondering about the details on the
version number referenced further down this commit message.
* Asana link: https://app.asana.com/0/529741067494252/658412373801330

This change addresses the need by:

* Upgrading Hackney version to ~> 1.12
* Upgrading HTTPoison to ~> 1.1.1
* Editing tests per PhoenixHTML 2.11 changes. For more info on these
change please see the PhoenixHTML PR:
https://github.com/phoenixframework/phoenix_html/pull/183