Match against stops in schedule for CR alerts
Why:

* Chris Hayes reported a bug in which he received an irrelevant alert
for a commuter rail subscription. The alert's informed entity didn't
include a stop id and it's trip didn't have a scheduled stop at the
subscription's origin. He has subscriptions from `Grafton -> Back Bay`
and `Worcester -> Back Bay` but the alert was for trip 591 which doesn't
stop at Grafton or Worcester. This notification was sent because
currently we don't consider the scheduled stops for a given trip when
figuring out if there should be a "stop match". So, because the alert
didn't include a stop and everything else matched, the notification was
sent.
* Asana link: https://app.asana.com/0/529741067494252/668866288927880

This change addresses the need by:

* Editing `InformedEntityFilter.stop_match?/1` to try to match against
the stops in a schedule when no stop is provided in the alert's informed
entity.
* Editing `AlertParser` to fetch the trip's schedule and add it to the
alert's informed entity. This is necessary for the changes made to
`InformedEntityFilter` to work. Note that we only do this for an
alert's informed_entity if it doesn't include a stop_id. This is because
if the stop_id is provided, we don't need the schedule to infer the
affected stops from when the alert passes through the
`InformedEntityFilter`.