Fix ServiceInfoCache not starting
Why:

* `ServiceInfoCache` is not starting because the API is not returning a
schedule for ferry ID "Boat-F4-1000-Long-Weekend" from the `/schedules`
endpoint. For more details please see the Asana ticket.
* Asana link: https://app.asana.com/0/529741067494252/673336152910622

This change addresses the need by:

* Editing `ServiceInfoCache.fetch_service_info(:ferry_general_ids)` to
not crash if the API can't find a schedule for a given trip ID. In these
cases we simply nil the `origin_id` and `departure_time` (for the first
scheduled event in a given schedule) that is used to create the
"generalized trip id".
* Deleting `ferry_mapper_test.exs` since it expects all trip ID's to
return a schedule. `FerryMapper` is not being used anymore so it's ok to
remove the test for it. Added an asana ticket to do so on a separate PR,
to be able to get this fix on prod as fast as possible.