Improve commuter rail trip matching
Why:

* Gene reported that he received an alert for a trip that departed his
subscription's origin after his subscription's start time. We'd like to
not send notifications in scenarios like this one.
* Asana link: https://app.asana.com/0/529741067494252/693372799954556

This change addresses the need by:

* Editing `InformedEntityFilter` by adding a new check called
`trip_match?`. This allows us to check if a given trip's departure time,
for the subscription's origin, is before the subscription's start time.
If it is, we don't have a "trip match".
* Editing the `facilities_alerts.json` ex_vcr cassette by updating the
`CR-Weekday-Fall-17-773` to `CR-Weekday-Spring-18-773`. Without this,
the newly added integration tests (in `MatchingTest`) won't work because
`CR-Weekday-Fall-17-773` doesn't exist in the `ServiceInfoCache`
anymore. This highlights a possible improvement opportunity for us to
consider if this becomes too painful: instead of having to manually
maintain cassettes like `facilities_alerts.json`, maybe we could set
things up in a way that would allow us to erase obsolete cassettes while
letting `ExVCR` create a new one for us when needed (as the
schedules/API responses change and so on).