Update the landing page design
Why: Add more context to our landing page to explain to users what our
service is and why they might want to sign up for it. Additionally,
clarify our relationship to the old GovDelivery system.

Note: the spec calls for the form field labels to slide in on focus. I
have pushed that piece off to a future commit for the purpose of getting
this out ASAP.