Fix text replacement bug (when hour is 12)
Why:

* While we were looking into a reported bug about the text replacement
feature not working on SMS, we found a bug in the `TextReplacement`
module that caused it to crash. The bug was that with a time of 12pm,
the logic incorrectly parses it as hour 24. Similarly with a time
of 12am, the logic in place incorrectly parses it as hour 12. For 12pm,
we should parse it as hour 12 and for 12am we should parse it as hour 0.
* Asana link: https://app.asana.com/0/529741067494252/706446575079957

This change addresses the need by:

* Editing `TextReplacement.parse_time/1` to parse 12pm as hour 12, and
12am as hour 0.