Improves InformedEntityFilter
Why:

* The `InformedEntityFilter` runs the given data through a set of
functions to determine if the given subscription should be sent a
notification for the given alert. Some of these functions are:
`trip_match?`, `route_type_match?`, and `direction_id_match?` among
others. If any of these functions don't "match" we don't have to run the
rest of them. Right now we run every single function, regardless of
their results. Changing this would allow us to shave off a bit of
matching time.
* Asana link: https://app.asana.com/0/529741067494252/709841296254769

This change addresses the need by:

* Editing `InformedEntityFilter` for it to work as described above.