SubscriptionFilterEngine async alert processing
Why:

* Performance; to reduce matching time.
* Asana link: https://app.asana.com/0/529741067494252/711047520528009

This change addressed the need by:

* Editing the `SubscriptionFilterEngine` to process alerts concurrently.

-------------------------------------------------------------------

Matching time measurements:

* Before (not concurrently with `ordered: true`):

    7679
    8554
    8877
    7779
    8046

* After (concurrently with `ordered: true`):

    5194
    5674
    4739
    4698
    5738

* After (concurrently with `ordered: false`):

    4690
    4738
    5507
    5526
    4789

-------------------------------------------------------------------

Opportunity for performance improvement:

Not accumulating the notifications and returning them as the output of
`schedule_all_notifications/1` will also shave some time. I ran some
numbers to test this but this commit doesn't include this change because
it breaks a decent number of tests. We can make this change on a
separate commit if we decide we want it. Here's the
difference vs. the matching times above:

* with `ordered: true`:

  4720
  4827
  4798
  4878
  5061

It should be faster with `ordered: false` but it's late now :P so we
can measure again if we allow time for this in the coming days.

------------------------------------------------------------------

Regarding the time measurements above, they were done as follows:

* using an unchanging alerts feed (copy from prod)
* freezing datetime at June 14, 8AM in UTC
* using 9,564 subscriptions (randomly multiplied subscriptions from prod)
* locally (dev env)