Add user multiplier script
Why:

* When load testing we want an easy way to create a lot of users,
subscriptions, and trips in the DB. The thinking behind this script is:
let's say you'd like to run a load teat with 10,000 users. First, get a
copy of the production DB and load it to your local dev DB or similar.
Second, use this "user multiplier" script to grow your database to
10,000 users.
* Asana link: https://app.asana.com/0/529741067494252/711493345220205

This change addresses the need by:

* Creating user multiplier script.
* Editing `User` model to include trips association. Necessary for user
multiplier script to work.