Display the final leg's destination in the trip card
Why:
• Trip card only destination of first leg rather than the final leg.
• Asana link: https://app.asana.com/0/529741067494252/728169399208095

I previously made a change (48a5f1a0ab539c5a181ea067cc374dd580f4e389)
attempting to fix a different issue where the destination of the final
leg of the return trip was being shown, which was the same as the
origin. My fix at the time was to only use the initial leg's origin and
destination, but that was incomplete. Instead we want to look at all
legs in the first trip, but not those in the return trip.