Create a trip review card
Adding a visual confirmation of what we're storing for first trip and
return trip on the trip create and edit pages.

Why:  Particularly for users who have multi-leg roundtrips, it can be
unclear what we're actually storing for them as their subscription on
the backend. This also serves as confirmation for the stops that the
user has selected.

Asana ticket: https://app.asana.com/0/529741067494252/705823114976616