Update the alert email design
Why: The current alert emails are over-designed and repeat the subject of the email as the header, making it take longer to see the relevant information. Additionally the image-rich emails take longer to load. We want to revamp the design to be simpler and focus on the content of the message.

Asana ticket: https://app.asana.com/0/529741067494252/779434909651643