Use SES API instead of SMTP for email sending
This allows us to lean on ExAWS's automatic backoff-and-retry for error
handling as we currently do for SMS, instead of dropping emails after a
single failed attempt.

Since the Bamboo SES adapter depends on Bamboo 1.x and we were still on
0.x, this involves some package updates. Renaming of `EmailPreviewPlug`
and `assert_delivered_with` are breaking changes in Bamboo 1.0. We are
also able to upgrade to a released version of `ex_aws_sns` that includes
the commit we were pointing to on GitHub.