Improve logging around notification emails
Notification emails are now delivered in-process so we can access the
adapter return value and know whether the sending succeeded or failed.
This is the same approach we currently use for SMS delivery.

This is expected to break continuity with some Splunk graphs and alerts
that keyed off the "Notification Email" text; nothing really depended on
"SMS notification" since we seem to prefer Amazon's SNS logs.