Improve notification performance logging
The existing logging here had an issue where it relied on a calculated
"optimal time" to have delivered the notification, which often ended up
in the future, causing the measurements to be negative. This was partly
because there are many complexities in how we decide when to send a
notification, and the calculation could not account for all of them.

To instead provide a concrete "start" to the processing timeline, this
change records the time when we fetched the alert that eventually caused
a notification to be sent. We add this as a field on our in-memory-only
Alert struct, so we don't have to separately pass it all the way through
the data pipeline.

Notably, this can't tell us the time between an alert being created or
updated, and the app ingesting it. This is especially relevant given the
delays on processing older alerts. But looking at an alert's timestamp
doesn't always give a meaningful answer, since we can intentionally send
notifications many hours after an update (based on users' time windows).
This is what the "optimal time" was trying to get at.

Other new information being logged:

* The IDs of all resources involved, in case we need to track them down

* The type of notification, since we may e.g. consider the performance
  of initial/update notifications to be more important than reminders

* The time bucket (recent/older/oldest) of the alert-processing run that
  generated the notification, for similar reasons

* The severity of the alert — "high priority" alerts always generate
  immediate notifications, so with these we can actually use the alert
  timestamp as an indicator of true end-to-end processing time