Remove de-duplication at sending queue level
This seems intended to address some issue with duplicate notifications,
but it hasn't done anything for quite a while; the notifications being
added to the queue are always inserted database records, and thus have
unique IDs and timestamps even if their other attributes are identical.
The `uniq` call on every push just slows down the operation.