Increase notification workers from 20 to 40
This is to test whether the long request times we've been seeing with
SES scale up with the number of parallel requests, or whether they're
capped at a certain point.

This also removes the pop error logging, since it didn't seem useful and
resulted in very high log noise when the application was idle.