Revert increase of notification workers
This reverts the earlier increase from 20 to 40, as we discovered a
bottleneck elsewhere in the app that needs to be addressed first, and
it's causing GenServer timeouts that result in emails being lost.

This also reverts the Hackney tracing change, as the multi-line format
proved almost impossible to analyze using Splunk.