Remove single-process bottleneck in email sending
`MailerInterface` was a GenServer seemingly created to proxy calls from
the AlertProcessor app to the ConciergeSite app's mailer, to avoid the
former having an explicit dependency on the latter, which would create a
cyclic dependency error. Technically this meant all email notifications
had to go through that single process, but since it used `deliver_later`
to perform the actual delivery in a separate process per email, calls to
it returned almost instantly.

In 9e462ee we moved from SMTP to the SES API for email sending, and
changed the mailer call from `deliver_later` to `deliver_now` so we'd
have access to the result of the API call. This turned `MailerInterface`
into a true bottleneck, since it could only send a single email at a
time, canceling out the speed benefits of the notification worker pool.
This also slowed SMS sending, since messages could not be sent when all
available workers were "waiting in line" to use the mailer.

Since there's no apparent need for a GenServer here, in this change we
remove it and call the mailer directly, fetching the module from app
configuration to continue avoiding an explicit cyclic dependency. App
logic from `MailerInterface` is moved to `NotificationSender`, renamed
from `NotificationSmser` to reflect that it now handles email as well.

We also go a bit further and entirely remove the Bamboo dependency from
AlertProcessor, as it was only referenced in a few places.