Refactor: extract bulk notif. saving/enqueueing
Previously the function of "save and enqueue a list of notifications"
was duplicated between the `Scheduler` and `Reminders.Processor`, and
implemented slightly differently in each place. Since we want to add
batch-inserting of notifications to improve database performance, this
would have meant more duplication.

This adapts the `Scheduler` into a more generic `MassNotifier` module
that performs only the "bulk save and enqueue" task, and can be called
from both places. Since the only other thing `Scheduler` did was build
the notifications, and this is nearly a one-liner, it is pulled out to
the calling code (as was already the case in the reminder processor).