Send notifications in queue order
The previous implementation of the `SendingQueue` was in fact a stack,
causing notifications to be sent in first-in-last-out order. While the
order of notifications within a given "run" of the alert processor is
not meaningful, over consecutive runs this could have odd user-facing
effects, such as getting an "update" for an alert prior to getting the
alert itself.