Validate email addresses more strictly
The previous validation only checked for an `@` somewhere in the string,
and since we perform no confirmation of email addresses, this allowed a
number of users to register despite common typos like forgetting the TLD
or inserting whitespace somewhere. This meant they would not receive any
alerts, if their preferred communication mode was email.

With this change we use the same regex used by the `mail` library, which
is itself based on RFC2822.