Use `insert_all` to save notification batches
The previous approach of wrapping the individual saves in a transaction,
while faster in local testing, turned out to be slower on the production
database. With this change we insert each batch of notifications in one
SQL statement (or two, since NotificationSubscription records must also
be created), falling back on individual saving if this raises an error.

Since the return value of `save_and_enqueue` was never used except in
tests, we can gain an additional bit of speed by not using `RETURNING`
for the batch inserts and having the function only return `:ok`.